+++ src/app/global.css
@@ -27,36 +27,12 @@ body {
     rgb(var(--background-start-rgb));
 }
 
-@import url("keyframes.css"); 
-@import url('https://fonts.googleapis.com/css2?family=Itim&display=swap');
-
-.App {
-	display: flex;
-	text-align: center;
-	flex-direction: column;
-	min-height: 100vh;
-	font-family: 'Itim', cursive;
-		
-  }
-  
-
-.container{
-  margin: 84px 0px 0px 60px;
-}
-
 /*スライドショー-----------------------------------------------------------------*/
-.slide{
-  margin: 0;
-}
-.slide #mainimg{
-  
 
-}
 .slide img{
   width: 100%;
   height: 100%;
   display: block;
-
 
 }
 
@@ -101,10 +77,7 @@ body {
 }
 
 
-
 /*footer--------------------------------------------------------------------*/
-
-
 footer {
 	
 	font-size: 80%;	/*文字サイズ*/
@@ -139,7 +112,8 @@ footer ul li{
 /*contents---------------------------------------------------------*/
 
 #contents {
-	padding: 0px 0 50px;	/*上、左右、下へのコンテンツ内の余白*/
+	padding: 0px 0px 0px 0px;	/*上、左右、下へのコンテンツ内の余白*/
+	margin: 65px 0px 100px 65px;
 	position: relative;
 	animation-name: opa2;	/*keyframes.cssで使う@keyframesの指定*/
 	animation-duration: 0.5S;	/*アニメーションの実行時間*/
@@ -252,13 +226,12 @@ footer ul li{
 	border-top: 1.5px solid #ccc;		/*上の線の幅、線種、色*/
 	border-bottom: 1.5px solid #ccc;	/*下の線の幅、線種、色*/
 	height: 105px;
-
-
+	
 }
 
 /*メニュー１個あたりの設定*/
 #menubar ul li {
-	margin: 30px 0px;	/*メニュー同士の間隔*/
+	margin: 10px 0px;	/*メニュー同士の間隔*/
 	float: left;	/*左に回り込み*/
 	width: 23%;		/*幅。今回は５個なので、100÷5=20。*/
 	font-size: 20px;	/*文字サイズ*/
