--- src/app/keyframes.css
@@ -1,16 +1,0 @@
-@charset "utf-8";
-
-
-/*opa1（透明から着色状態に）
----------------------------------------------------------------------------*/
-@keyframes opa1 {
-	0% {opacity: 0;}
-	100% {opacity: 1;}
-}
-
-/*opa2（透明から着色状態に＆下からフェードイン）
----------------------------------------------------------------------------*/
-@keyframes opa2 {
-	0% {opacity: 0;top: 10px;}
-	100% {opacity: 1;top: 0px;}
-}
