+++ src/app/layout.js
@@ -1,4 +1,6 @@
 import './global.css'
+import Header from './components/Header'
+import Footer from './components/Footer'
 export const metadata = {
   title: 'Next.js',
   description: 'Generated by Next.js',
@@ -7,7 +9,11 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang='en'>
-      <body>{children}</body>
+      <body>      
+        <Header />
+          {children}
+        <Footer />
+      </body>
     </html>
   )
 }
