--- src/app/slide.css
@@ -1,59 +1,0 @@
-@charset "utf-8";
-
-
-/*CSSスライドショー設定
----------------------------------------------------------------------------*/
-/*１枚目*/
-@keyframes slide1 {
-	0% {opacity: 0;}
-	10% {opacity: 1;}
-	35% {opacity: 1;}
-	45% {opacity: 0;}
-	100% {opacity: 0;}
-}
-/*２枚目*/
-@keyframes slide2 {
-	0% {opacity: 0;}
-	25% {opacity: 0;}
-	35% {opacity: 1;}
-	60% {opacity: 1;}
-	70% {opacity: 0;}
-	100% {opacity: 0;}
-}
-/*３枚目*/
-@keyframes slide3 {
-	0% {opacity: 0;}
-	50% {opacity: 0;}
-	60% {opacity: 1;}
-	75% {opacity: 1;}
-	85% {opacity: 0;}
-	100% {opacity: 0;}
-}
-
-/*mainimg
----------------------------------------------------------------------------*/
-#mainimg {
-	clear: left;
-	width: 100%;
-	position: relative;
-}
-.slide1,.slide2,.slide3 {
-	animation-duration: 15s;	/*アニメーションを実行する時間。「s」は秒の事。*/
-	animation-iteration-count:infinite;
-	position: absolute;left:0px;top:0px;width: 100%;height: auto;
-	animation-fill-mode: both;	/*アニメーションの待機中は最初のキーフレームを維持、終了後は最後のキーフレームを維持。*/
-	animation-delay: 2s;		/*出現するタイミング（秒後）*/
-}
-.slide0 {
-	position: relative;width: 100%;height: auto;z-index: -1;
-}
-.slide1 {
-	animation-name: slide1;
-}
-.slide2 {
-	animation-name: slide2;
-}
-.slide3 {
-	animation-name: slide3;
-}
-
